Fixed bug for wavelet time bins
Before we were setting N = length of data, regardless of whether it is time or frequency domain.
This is only correct if we use a two-sided transform where the length of FFT = length in time domain.

For zero_padded signals (as they all should be, for speed), the rfft returns N/2 points. Hence, in order to get the
correct time bins, we need to double the data points if we take in a frequency series.

This was fine for the time domain, but incorrect for the frequency domain.

Ollie