Normalising constants, understood.
Matt's code is different from Cornish's code. For Matt's code to be consistent with our
frequency domain code, we require a normalising factor in front of the Meyer window
of the form $(Nf/2) \cdot \pi^{-1/2}$. On this specific commit, there are a load of
comments in the function phitilde_vec_norm indicating parts we need to understand.

The nice thing though is that analytically, for monochromatic signals, we now
have an expression for the wavelet coefficients $\omega_{nm} = A\sqrt{2N_{f}}$ for
n odd and m even. With the conventions above, we have verified this + checked the SNRs.

I'm now happy with this transformation code.